move puppet node graph methods to node graph file
